Correct narrowing during assignment (gcc 4.7.2 issues).

+++ src/XrdCms/XrdCmsTalk.cc
@@ -101,7 +101,7 @@ const char *XrdCmsTalk::Request(XrdLink *Link, XrdCms::CmsRRHdr &Hdr,
                                 char    *buff, int blen)
 {
    struct iovec ioV[2] = {{(char *)&Hdr, sizeof(Hdr)},
-                          {(char *)buff, blen}};
+                          {(char *)buff, (size_t)blen}};
 
    Hdr.datalen = htons(static_cast<unsigned short>(blen));
 
@@ -122,7 +122,7 @@ const char *XrdCmsTalk::Respond(XrdLink *Link, XrdCms::CmsRspCode rcode,
    CmsResponse Resp = {{0, rcode, 0,
                         htons(static_cast<unsigned short>(blen+ovhd))}, 0};
    struct iovec ioV[2] = {{(char *)&Resp, sizeof(Resp)},
-                          {         buff, blen}};
+                          {         buff, (size_t)blen}};
 
 // Send the actual data
 //
